LayoutTests/inspector/console/x-frame-options-message-expected.txt

CONSOLE MESSAGE: The X-Frame-Option 'deny' supplied in a <meta> element was ignored. X-Frame-Options may only be provided by an HTTP header sent with the document.


== Running test suite: Console.XFrameOptionsMessages
-- Running test case: XFrameOptionsDeny
{"targetType":"frame","messageText":"The X-Frame-Option 'deny' supplied in a <meta> element was ignored. X-Frame-Options may only be provided by an HTTP header sent with the document.","source":"security","level":"error"}
